backend/requirements.txt
fastapi==0.99.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.19
psycopg2-binary==2.9.6
boto3==2.7.0
python-dotenv==1.0.0
pydantic==1.10.12
celery[redis]==5.3.1
redis==4.5.5
Pillow==10.0.0
requests==2.31.0